feat: show usage status indicator
