+++ react/package.json
@@ -12,7 +12,7 @@
     "@types/react": "^18.2.23",
     "@types/react-dom": "^18.2.8",
     "ahooks": "^3.7.8",
-    "antd": "^5.10.1",
+    "antd": "^5.13.2",
     "dayjs": "^1.11.10",
     "i18next": "^23.5.1",
     "i18next-http-backend": "^2.2.2",
