+++ react/src/components/StorageSelect.css
@@ -1,0 +1,23 @@
+.host-status-indicator {
+  height: 16px;
+  padding-left: 8px;
+  padding-right: 8px;
+  border-radius: 8px;
+  font-size: 12px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: #ffffff;
+}
+
+.host-status-indicator.adequate {
+  background-color: rgba(58, 178, 97, 1);
+}
+
+.host-status-indicator.caution {
+  background-color: rgb(223, 179, 23);
+}
+
+.host-status-indicator.insufficient {
+  background-color: #ef5350;
+}
